+++ src/App/Fossa/Analyze.hs
@@ -19,7 +19,7 @@ import App.Fossa.Analyze.Discover (
   discoverFuncs,
  )
 import App.Fossa.Analyze.Filter (
-  CountedResult (FilteredAll, FoundSome, NoneDiscovered),
+  CountedResult (..),
   checkForEmptyUpload,
  )
 import App.Fossa.Analyze.GraphMangler (graphingToGraph)
@@ -32,7 +32,7 @@ import App.Fossa.Analyze.Types (
   DiscoveredProjectIdentifier (..),
   DiscoveredProjectScan (..),
  )
-import App.Fossa.Analyze.Upload (uploadSuccessfulAnalysis)
+import App.Fossa.Analyze.Upload (mergeSourceAndLicenseUnits, uploadSuccessfulAnalysis)
 import App.Fossa.BinaryDeps (analyzeBinaryDeps)
 import App.Fossa.Config.Analyze (
   AnalyzeCliOpts,
@@ -44,12 +44,11 @@ import App.Fossa.Config.Analyze (
   IncludeAll (IncludeAll),
   NoDiscoveryExclusion (NoDiscoveryExclusion),
   ScanDestination (..),
-  StandardAnalyzeConfig (severity),
   UnpackArchives (UnpackArchives),
  )
 import App.Fossa.Config.Analyze qualified as Config
+import App.Fossa.FirstPartyScan (runFirstPartyScan)
 import App.Fossa.ManualDeps (analyzeFossaDepsFile)
-import App.Fossa.Monorepo (monorepoScan)
 import App.Fossa.Subcommand (SubCommand)
 import App.Fossa.VSI.DynLinked (analyzeDynamicLinkedDeps)
 import App.Fossa.VSI.IAT.AssertRevisionBinaries (assertRevisionBinaries)
@@ -57,6 +56,7 @@ import App.Fossa.VSI.Types qualified as VSI
 import App.Fossa.VSIDeps (analyzeVSIDeps)
 import App.Types (
   BaseDir (..),
+  FirstPartyScansFlag (..),
   OverrideDynamicAnalysisBinary,
   ProjectRevision (..),
  )
@@ -90,6 +90,7 @@ import Data.Aeson qualified as Aeson
 import Data.ByteString.Lazy qualified as BL
 import Data.Flag (Flag, fromFlag)
 import Data.Foldable (traverse_)
+import Data.List.NonEmpty qualified as NE
 import Data.Maybe (mapMaybe)
 import Data.String.Conversion (decodeUtf8, toText)
 import Data.Text.Extra (showT)
@@ -119,7 +120,7 @@ import Prettyprinter.Render.Terminal (
   color,
  )
 import Srclib.Converter qualified as Srclib
-import Srclib.Types (Locator, SourceUnit)
+import Srclib.Types (LicenseSourceUnit, Locator, SourceUnit, sourceUnitToFullSourceUnit)
 import Types (DiscoveredProject (..), FoundTargets)
 
 debugBundlePath :: FilePath
@@ -139,9 +140,7 @@ dispatch ::
   ) =>
   AnalyzeConfig ->
   m ()
-dispatch = \case
-  Monorepo cfg -> monorepoScan cfg
-  Standard cfg -> void $ analyzeMain cfg
+dispatch cfg = void $ analyzeMain cfg
 
 -- This is just a handler for the Debug effect.
 -- The real logic is in the inner analyze
@@ -154,7 +153,7 @@ analyzeMain ::
   , Has ReadFS sig m
   , Has Telemetry sig m
   ) =>
-  StandardAnalyzeConfig ->
+  AnalyzeConfig ->
   m Aeson.Value
 analyzeMain cfg = case Config.severity cfg of
   SevDebug -> do
@@ -243,7 +242,7 @@ analyze ::
   , Has ReadFS sig m
   , Has Telemetry sig m
   ) =>
-  StandardAnalyzeConfig ->
+  AnalyzeConfig ->
   m Aeson.Value
 analyze cfg = Diag.context "fossa-analyze" $ do
   capabilities <- sendIO getNumCapabilities
@@ -295,6 +294,15 @@ analyze cfg = Diag.context "fossa-analyze" $ do
       additionalSourceUnits = mapMaybe (join . resultToMaybe) [manualSrcUnits, vsiResults, binarySearchResults, dynamicLinkedResults]
   traverse_ (Diag.flushLogs SevError SevDebug) [vsiResults, binarySearchResults, manualSrcUnits, dynamicLinkedResults]
 
+  maybeFirstPartyScanResults <-
+    Diag.errorBoundaryIO . diagToDebug $
+      if firstPartyScansFlag cfg == FirstPartyScansOffFromFlag
+        then do
+          logInfo "first party scans forced off by the --experimental-turn-off-scans  flag. Skipping first party scans"
+          pure Nothing
+        else Diag.context "first-party-scans" . runStickyLogger SevInfo $ runFirstPartyScan basedir maybeApiOpts cfg
+  let firstPartyScanResults = join . resultToMaybe $ maybeFirstPartyScanResults
+
   let discoveryFilters = if fromFlag NoDiscoveryExclusion noDiscoveryExclusion then mempty else filters
   (projectScans, ()) <-
     Diag.context "discovery/analysis tasks"
@@ -330,19 +338,22 @@ analyze cfg = Diag.context "fossa-analyze" $ do
   renderScanSummary (severity cfg) maybeEndpointAppVersion analysisResult $ Config.filterSet cfg
 
   -- Need to check if vendored is empty as well, even if its a boolean that vendoredDeps exist
-  let result = buildResult includeAll additionalSourceUnits filteredProjects
-  case checkForEmptyUpload includeAll projectResults filteredProjects additionalSourceUnits of
+  let result = buildResult includeAll additionalSourceUnits filteredProjects firstPartyScanResults
+  case checkForEmptyUpload includeAll projectResults filteredProjects additionalSourceUnits firstPartyScanResults of
     NoneDiscovered -> Diag.fatal ErrNoProjectsDiscovered
     FilteredAll -> Diag.fatal ErrFilteredAllProjects
-    FoundSome sourceUnits -> case destination of
-      OutputStdout -> logStdout . decodeUtf8 $ Aeson.encode result
-      UploadScan apiOpts metadata ->
-        Diag.context "upload-results"
-          . runFossaApiClient apiOpts
-          $ do
-            locator <- uploadSuccessfulAnalysis (BaseDir basedir) metadata jsonOutput revision sourceUnits
-            doAssertRevisionBinaries iatAssertion locator
+    CountedScanUnits scanUnits -> doUpload result iatAssertion destination basedir jsonOutput revision scanUnits
   pure result
+  where
+    doUpload result iatAssertion destination basedir jsonOutput revision scanUnits =
+      case destination of
+        OutputStdout -> logStdout . decodeUtf8 $ Aeson.encode result
+        UploadScan apiOpts metadata ->
+          Diag.context "upload-results"
+            . runFossaApiClient apiOpts
+            $ do
+              locator <- uploadSuccessfulAnalysis (BaseDir basedir) metadata jsonOutput revision scanUnits
+              doAssertRevisionBinaries iatAssertion locator
 
 toProjectResult :: DiscoveredProjectScan -> Maybe ProjectResult
 toProjectResult (SkippedDueToProvidedFilter _) = Nothing
@@ -449,13 +460,17 @@ instance Diag.ToDiagnostic AnalyzeError where
       , ""
       ]
 
-buildResult :: Flag IncludeAll -> [SourceUnit] -> [ProjectResult] -> Aeson.Value
-buildResult includeAll srcUnits projects =
+buildResult :: Flag IncludeAll -> [SourceUnit] -> [ProjectResult] -> Maybe LicenseSourceUnit -> Aeson.Value
+buildResult includeAll srcUnits projects firstPartyScanResults =
   Aeson.object
     [ "projects" .= map buildProject projects
-    , "sourceUnits" .= finalSourceUnits
+    , "sourceUnits" .= mergedUnits
     ]
   where
+    mergedUnits = case firstPartyScanResults of
+      Nothing -> map sourceUnitToFullSourceUnit finalSourceUnits
+      Just licenseUnits -> do
+        NE.toList $ mergeSourceAndLicenseUnits finalSourceUnits licenseUnits
     finalSourceUnits = srcUnits ++ scannedUnits
     scannedUnits = map (Srclib.toSourceUnit (fromFlag IncludeAll includeAll)) projects