+++ src/App/Fossa/EmbeddedBinary.hs
@@ -7,7 +7,6 @@ module App.Fossa.EmbeddedBinary (
   ThemisIndex,
   ThemisBins (..),
   toPath,
-  withWigginsBinary,
   withThemisAndIndex,
   withBerkeleyBinary,
   allBins,
@@ -48,8 +47,7 @@ import Path.IO (
 import Prelude hiding (writeFile)
 
 data PackagedBinary
-  = Wiggins
-  | Themis
+  = Themis
   | ThemisIndex
   | BerkeleyDB
   deriving (Show, Eq, Enum, Bounded)
@@ -98,13 +96,6 @@ extractThemisFiles = do
   _ <- sendIO $ BL.writeFile (toString $ toPath decompressedThemisIndex) (Lzma.decompress $ toLazy embeddedBinaryThemisIndex)
   pure $ ThemisBins themisActual $ applyTag @ThemisIndex decompressedThemisIndex
 
-withWigginsBinary ::
-  ( Has (Lift IO) sig m
-  ) =>
-  (BinaryPaths -> m c) ->
-  m c
-withWigginsBinary = withEmbeddedBinary Wiggins
-
 withBerkeleyBinary ::
   ( Has (Lift IO) sig m
   ) =>
@@ -140,7 +131,6 @@ dumpEmbeddedBinary dir bin = writeBinary path bin
 
 writeBinary :: (Has (Lift IO) sig m) => Path Abs File -> PackagedBinary -> m ()
 writeBinary dest bin = sendIO . writeExecutable dest $ case bin of
-  Wiggins -> embeddedBinaryWiggins
   Themis -> embeddedBinaryThemis
   ThemisIndex -> embeddedBinaryThemisIndex
   BerkeleyDB -> embeddedBinaryBerkeleyDB
@@ -153,9 +143,6 @@ writeExecutable path content = do
 
 extractedPath :: PackagedBinary -> Path Rel File
 extractedPath bin = case bin of
-  -- Rename wiggins upon local extraction so that we can provide a better status line to users during the VSI strategy.
-  -- Users don't know what "wiggins" is, but they explicitly enable the VSI plugin, so this is more intuitive.
-  Wiggins -> $(mkRelFile "vsi-plugin")
   Themis -> $(mkRelFile "themis-cli")
   ThemisIndex -> $(mkRelFile "index.gob.xz")
   BerkeleyDB -> $(mkRelFile "berkeleydb-plugin")
@@ -187,9 +174,6 @@ makeExecutable path = do
 -- built binaries of the appropriate architecture.
 -- The below functions are expected to warn since the vendor-bins directory is typically populated in CI.
 -- If you wish to run these on your local system, populate these binaries via `vendor_download.sh`.
-embeddedBinaryWiggins :: ByteString
-embeddedBinaryWiggins = $(embedFileIfExists "vendor-bins/wiggins")
-
 embeddedBinaryThemis :: ByteString
 embeddedBinaryThemis = $(embedFileIfExists "vendor-bins/themis-cli")